move query to chain state provider
